OpenGL/Triangulate: don't use GLushort for count
Use the default unsigned integer, the fastest data type the CPU has.

+++ src/Screen/OpenGL/Triangulate.cpp
@@ -290,7 +290,7 @@ TriangleToStrip(GLushort *triangles, unsigned index_count,
     return 0;
 
   // count the number of occurrences for each vertex
-  GLushort *vcount = new GLushort[vertex_count]();
+  unsigned *vcount = new unsigned[vertex_count]();
   GLushort *t = triangles;
   GLushort *v;
   const GLushort * const t_end = triangles + index_count;
